Refactored UI code
