+++ app/pages/ExplorePage/components/ImageModal/index.ts
@@ -1,0 +1,1 @@
+export { default as ImageModal } from "./ImageModal";
