+++ app/types/Images.d.ts
@@ -1,7 +1,25 @@
+export type Comment = {
+  id: string;
+  message: string;
+  createdAt: string;
+  updatedAt: string;
+  user: {
+    id: string;
+    username: string;
+  };
+  parentId: string;
+};
+
 export type ImageType = {
   id: string;
   prompt: string;
+  title?: string;
   url: string;
   createdAt: Date;
-  createdBy: string;
+  // createdBy: string;
+  user: {
+    id: string;
+    username: string;
+  };
+  comments: Comment[];
 };
